# example/openapi.yaml
openapi: 3.0.3

info:
  title: Task Service API
  version: 1.0.0
  description: >
    Example task and project management API used to test API Observer
    with contract validation and generated traffic.

servers:
  - url: http://localhost:8080

paths:
  /health:
    get:
      operationId: getHealth
      summary: Health check
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                properties:
                  status:
                    type: string
                    enum:
                      - ok

  /users:
    get:
      operationId: listUsers
      summary: List users
      parameters:
        - name: role
          in: query
          required: false
          schema:
            type: string
            enum:
              - admin
              - user
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

    post:
      operationId: createUser
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

        "400":
          $ref: "#/components/responses/BadRequest"

        "409":
          description: User email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "422":
          $ref: "#/components/responses/ValidationError"

  /users/{id}:
    get:
      operationId: getUser
      summary: Get user by ID
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

        "400":
          $ref: "#/components/responses/BadRequest"

        "404":
          $ref: "#/components/responses/NotFound"

  /projects:
    get:
      operationId: listProjects
      summary: List projects
      parameters:
        - name: owner_id
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

        "400":
          $ref: "#/components/responses/BadRequest"

    post:
      operationId: createProject
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

        "400":
          $ref: "#/components/responses/BadRequest"

        "422":
          $ref: "#/components/responses/ValidationError"

  /projects/{id}:
    get:
      operationId: getProject
      summary: Get project by ID
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: Project found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

        "400":
          $ref: "#/components/responses/BadRequest"

        "404":
          $ref: "#/components/responses/NotFound"

  /projects/{projectID}/tasks:
    get:
      operationId: listProjectTasks
      summary: List tasks for a project
      parameters:
        - $ref: "#/components/parameters/ProjectPathID"

        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TaskStatus"

        - name: priority
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TaskPriority"

      responses:
        "200":
          description: List of project tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

        "400":
          $ref: "#/components/responses/BadRequest"

        "404":
          $ref: "#/components/responses/NotFound"

    post:
      operationId: createTask
      summary: Create task in project
      parameters:
        - $ref: "#/components/parameters/ProjectPathID"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"

      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

        "400":
          $ref: "#/components/responses/BadRequest"

        "404":
          $ref: "#/components/responses/NotFound"

        "422":
          $ref: "#/components/responses/ValidationError"

  /tasks/{id}:
    get:
      operationId: getTask
      summary: Get task by ID
      parameters:
        - $ref: "#/components/parameters/TaskID"
      responses:
        "200":
          description: Task found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

        "400":
          $ref: "#/components/responses/BadRequest"

        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      operationId: updateTask
      summary: Update task
      parameters:
        - $ref: "#/components/parameters/TaskID"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"

      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

        "400":
          $ref: "#/components/responses/BadRequest"

        "404":
          $ref: "#/components/responses/NotFound"

        "422":
          $ref: "#/components/responses/ValidationError"

    delete:
      operationId: deleteTask
      summary: Delete task
      parameters:
        - $ref: "#/components/parameters/TaskID"
      responses:
        "204":
          description: Task deleted successfully

        "400":
          $ref: "#/components/responses/BadRequest"

        "404":
          $ref: "#/components/responses/NotFound"

  /admin/stats:
    get:
      operationId: getAdminStats
      summary: Get administrative statistics
      parameters:
        - name: X-Admin-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 1
          example: observer-test-admin

      responses:
        "200":
          description: Administrative statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

        "401":
          description: Invalid admin credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  parameters:
    UserID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1

    ProjectID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1

    ProjectPathID:
      name: projectID
      in: path
      required: true
      schema:
        type: integer
        minimum: 1

    TaskID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: string
          minLength: 1

    UserRole:
      type: string
      enum:
        - admin
        - user

    User:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - email
        - role
        - created_at
      properties:
        id:
          type: integer
          minimum: 1

        name:
          type: string
          minLength: 1

        email:
          type: string
          format: email

        role:
          $ref: "#/components/schemas/UserRole"

        created_at:
          type: string
          format: date-time

    CreateUserRequest:
      type: object
      additionalProperties: false
      required:
        - name
        - email
        - role
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100

        email:
          type: string
          format: email
          maxLength: 254

        role:
          $ref: "#/components/schemas/UserRole"

    Project:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - owner_id
        - created_at
      properties:
        id:
          type: integer
          minimum: 1

        name:
          type: string
          minLength: 1

        description:
          type: string

        owner_id:
          type: integer
          minimum: 1

        created_at:
          type: string
          format: date-time

    CreateProjectRequest:
      type: object
      additionalProperties: false
      required:
        - name
        - owner_id
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 150

        description:
          type: string
          maxLength: 1000

        owner_id:
          type: integer
          minimum: 1

    TaskStatus:
      type: string
      enum:
        - todo
        - in_progress
        - done

    TaskPriority:
      type: string
      enum:
        - low
        - medium
        - high

    Task:
      type: object
      additionalProperties: false
      required:
        - id
        - project_id
        - title
        - status
        - priority
        - created_at
        - updated_at

      properties:
        id:
          type: integer
          minimum: 1

        project_id:
          type: integer
          minimum: 1

        title:
          type: string
          minLength: 1

        description:
          type: string

        status:
          $ref: "#/components/schemas/TaskStatus"

        priority:
          $ref: "#/components/schemas/TaskPriority"

        assignee_id:
          type: integer
          minimum: 1
          nullable: true

        due_date:
          type: string
          format: date-time
          nullable: true

        created_at:
          type: string
          format: date-time

        updated_at:
          type: string
          format: date-time

    CreateTaskRequest:
      type: object
      additionalProperties: false
      required:
        - title
        - priority

      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200

        description:
          type: string
          maxLength: 2000

        priority:
          $ref: "#/components/schemas/TaskPriority"

        assignee_id:
          type: integer
          minimum: 1
          nullable: true

        due_date:
          type: string
          format: date-time

    UpdateTaskRequest:
      type: object
      additionalProperties: false
      minProperties: 1

      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200

        description:
          type: string
          maxLength: 2000

        status:
          $ref: "#/components/schemas/TaskStatus"

        priority:
          $ref: "#/components/schemas/TaskPriority"

        assignee_id:
          type: integer
          minimum: 1
          nullable: true

        due_date:
          type: string
          format: date-time
          nullable: true

    AdminStats:
      type: object
      additionalProperties: false
      required:
        - users
        - projects
        - tasks

      properties:
        users:
          type: integer
          minimum: 0

        projects:
          type: integer
          minimum: 0

        tasks:
          type: integer
          minimum: 0